Changes in default launcher chooser
