+++ app/src/main/java/app/olauncher/data/Constants.kt
@@ -68,6 +68,7 @@ object Constants {
     const val FLAG_SET_CALENDAR_APP = 14
 
     const val REQUEST_CODE_ENABLE_ADMIN = 666
+    const val REQUEST_CODE_LAUNCHER_SELECTOR = 678
 
     const val HINT_RATE_US = 15
     const val HINT_SHARE = 22
