+++ app/src/main/java/app/olauncher/helper/Extensions.kt
@@ -1,8 +1,15 @@
 package app.olauncher.helper
 
+import android.app.Activity
+import android.app.role.RoleManager
+import android.content.ComponentName
 import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
 import android.view.View
 import android.view.inputmethod.InputMethodManager
+import androidx.annotation.RequiresApi
 
 fun View.hideKeyboard() {
     this.clearFocus()
@@ -17,4 +24,36 @@ fun View.showKeyboard(show: Boolean = true) {
             val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
             imm.toggleSoftInput(InputMethodManager.SHOW_FORCED, InputMethodManager.HIDE_IMPLICIT_ONLY)
         }, 100)
+}
+
+
+@RequiresApi(Build.VERSION_CODES.Q)
+fun Activity.showLauncherSelector(requestCode: Int) {
+    val roleManager = getSystemService(Context.ROLE_SERVICE) as RoleManager
+    if (roleManager.isRoleAvailable(RoleManager.ROLE_HOME)) {
+        val intent = roleManager.createRequestRoleIntent(RoleManager.ROLE_HOME)
+        startActivityForResult(intent, requestCode)
+    } else
+        resetDefaultLauncher()
+}
+
+fun Context.resetDefaultLauncher() {
+    try {
+        val componentName = ComponentName(this, FakeHomeActivity::class.java)
+        packageManager.setComponentEnabledSetting(
+            componentName,
+            PackageManager.COMPONENT_ENABLED_STATE_ENABLED,
+            PackageManager.DONT_KILL_APP
+        )
+        val selector = Intent(Intent.ACTION_MAIN)
+        selector.addCategory(Intent.CATEGORY_HOME)
+        startActivity(selector)
+        packageManager.setComponentEnabledSetting(
+            componentName,
+            PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
+            PackageManager.DONT_KILL_APP
+        )
+    } catch (e: Exception) {
+        e.printStackTrace()
+    }
 }
